Handle removing a non-existent pet

+++ src/main/java/org/springframework/samples/petclinic/service/ClinicServiceImpl.java
@@ -18,6 +18,7 @@ package org.springframework.samples.petclinic.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.Cacheable;
+import org.springframework.orm.ObjectRetrievalFailureException;
 import org.springframework.samples.petclinic.model.*;
 import org.springframework.samples.petclinic.repository.*;
 import org.springframework.stereotype.Service;
@@ -144,6 +145,11 @@ public class ClinicServiceImpl implements ClinicService {
     }
 
     public void removePetById(int petId) {
+        Pet pet = petRepository.findById(petId);
+        if (pet == null) {
+            throw new ObjectRetrievalFailureException("Pet not found", ObjectRetrievalFailureException.class);
+        }
+
         petRepository.removePet(petId);
     }
